add switch itowns view
