+++ client/assets/config/config_editor.json
@@ -49,6 +49,9 @@
       "static_object": {
         "path": "./assets/localScripts/StaticObject.js"
       },
+      "switch_itowns": {
+        "path": "./assets/localScripts/SwitchItowns.js"
+      },
       "camera": {
         "path": "./assets/localScripts/Camera.js"
       },
