+++ client/assets/config/config_features.json
@@ -1,3 +1,3 @@
 {
-  "editor": false
+  "editor": true
 }
