+++ client/assets/config/config_game.json
@@ -57,6 +57,9 @@
       "post_it": {
         "path": "./assets/localScripts/PostIt.js"
       },
+      "switch_itowns": {
+        "path": "./assets/localScripts/SwitchItowns.js"
+      },
       "camera": {
         "path": "./assets/localScripts/Camera.js"
       },
