+++ client/assets/localScripts/AvatarController.js
@@ -16,14 +16,9 @@ module.exports = class Controller {
 
     udviz = udvizBundle;
     Game = udviz.Game;
-    itowns = udviz.itowns;
 
     //Avatar controller
     this.avatarControllerMode = false;
-
-    //buffer DEBUG
-    // this.itownsCamPos = null;
-    // this.itownsCamQuat = null;
   }
 
   init() {
@@ -32,8 +27,6 @@ module.exports = class Controller {
     const camera = gameView.getCamera();
     const manager = gameView.getInputManager();
 
-    //DEBUG inputs
-    const _this = this;
     manager.addKeyInput('p', 'keydown', function () {
       console.log('DEBUG');
       console.log('Gameview ', gameView);
@@ -248,130 +241,5 @@ module.exports = class Controller {
 
   getAvatarControllerMode() {
     return this.avatarControllerMode;
-  }
-
-  //DEBUG
-  initSwitchItownsController(localCtx) {
-    const _this = this;
-
-    const gameView = localCtx.getGameView();
-    const div = gameView.getRenderer().domElement;
-    const camera = gameView.getCamera();
-    const manager = gameView.getInputManager();
-    const Routine = Game.Components.Routine;
-    const Command = Game.Command;
-
-    //SWITCH CONTROLS
-    const view = gameView.getItownsView();
-    if (view) {
-      manager.addKeyInput('a', 'keydown', function () {
-        if (_this.avatarCameraman.hasRoutine()) return; //already routine
-
-        const duration = 2000;
-        let currentTime = 0;
-        const camera = _this.avatarCameraman.getCamera();
-
-        let startPos = camera.position.clone();
-        let startQuat = camera.quaternion.clone();
-
-        if (view.controls) {
-          //record
-          const camera = _this.avatarCameraman.getCamera();
-          _this.itownsCamPos.set(
-            camera.position.x,
-            camera.position.y,
-            camera.position.z
-          );
-          _this.itownsCamQuat.setFromEuler(camera.rotation);
-
-          _this.avatarCameraman.addRoutine(
-            new Routine(
-              function (dt) {
-                const t = _this.avatarCameraman.computeTransformTarget();
-
-                //no avatar yet
-                if (!t) return false;
-
-                currentTime += dt;
-                let ratio = currentTime / duration;
-                ratio = Math.min(Math.max(0, ratio), 1);
-
-                const p = t.position.lerp(startPos, 1 - ratio);
-                const q = t.quaternion.slerp(startQuat, 1 - ratio);
-
-                camera.position.copy(p);
-                camera.quaternion.copy(q);
-
-                camera.updateProjectionMatrix();
-
-                view.notifyChange(); //trigger camera event
-
-                return ratio >= 1;
-              },
-              function () {
-                view.controls.dispose();
-                view.controls = null;
-              }
-            )
-          );
-        } else {
-          if (!_this.itownsCamPos && !_this.itownsCamQuat) {
-            //first time camera in sky
-
-            const currentPosition = new Game.THREE.Vector3().copy(
-              _this.avatarCameraman.getCamera().position
-            );
-
-            //200 meters up
-            const endPosition = new Game.THREE.Vector3(0, 0, 200).add(
-              currentPosition
-            );
-
-            //look down
-            const endQuaternion = new Game.THREE.Quaternion().setFromEuler(
-              new Game.THREE.Euler(0, 0, 0)
-            );
-
-            _this.itownsCamPos = endPosition;
-            _this.itownsCamQuat = endQuaternion;
-          }
-
-          _this.avatarCameraman.addRoutine(
-            new Routine(
-              function (dt) {
-                currentTime += dt;
-                let ratio = currentTime / duration;
-                ratio = Math.min(Math.max(0, ratio), 1);
-
-                const p = _this.itownsCamPos.clone().lerp(startPos, 1 - ratio);
-                const q = _this.itownsCamQuat
-                  .clone()
-                  .slerp(startQuat, 1 - ratio);
-
-                camera.position.copy(p);
-                camera.quaternion.copy(q);
-
-                camera.updateProjectionMatrix();
-
-                view.notifyChange(); //trigger camera event
-
-                return ratio >= 1;
-              },
-              function () {
-                manager.setPointerLock(false);
-
-                //creating controls like put it in _this.view.controls
-                const c = new itowns.PlanarControls(view, {
-                  handleCollision: false,
-                  focusOnMouseOver: false, //TODO itowns bug not working
-                  focusOnMouseClick: false,
-                  zoomFactor: 0.9, //TODO working ?
-                });
-              }
-            )
-          );
-        }
-      });
-    }
   }
 };
