+++ client/assets/worlds/Flying_Campus.json
@@ -24,6 +24,7 @@
         "uuid": "4C2E2FEF-0AD0-4EAA-BB4C-CB385AA19B3F",
         "idScripts": [
           "avatar_controller",
+          "switch_itowns",
           "camera",
           "zeppelin_controller",
           "static_object",
